LayoutTests/imported/w3c/web-platform-tests/html/browsers/history/the-location-interface/location-ancestor-origins-inactive-document.sub-expected.txt


PASS location.ancestorOrigins returns empty list after iframe is removed and referenced Location's relevant document is null
PASS location.ancestorOrigins returns empty list when iframe navigated away and referenced Location's relevant document is null
